requirements.txt
streamlit
openai>=1.0.0
faiss-cpu
pypdf
tiktoken
python-dotenv
numpy